Move ClientError, ServerError, and NetworkError from data -> jsonapi
